Implement alias zones
Usually, zone file names are directly derived from zone names.
However, there are times when one might want them to differ.
For backward compatibility, it became too cumbersome to redefine
'zones', as a new 'aliaszones' is created.

This is useful for more complex BIND setups, with autogenerated
reverse zones, where the generator locate the zones to generate
entries for by zone file name, but the desired domain for the PTR
records is different.

Example:

    [% aliaszones = {
           'net4.example.com' = '192/27.0.168.192.in-addr.arpa',
           'net6.example.com' = 'd.c.b.a.d.c.b.a.d.c.b.a.d.c.b.a.ip6.arpa'
       };
       masters = 'primary.example.com';
       slaves = 'secondary.example.com';
    -%]
    @               IN      SOA     [% host %]. hostmaster.example.com. (
                                    $SERIAL ; Serial
                                    [% default.soa.refresh %]       ; Refresh
                                    [% default.soa.retry   %]       ; Retry
                                    [% default.soa.expire  %]       ; Expire
                                    [% default.soa.minimum %]       ; Minimum
                                    )
    @               IN      NS      primary.example.com.
    @               IN      NS      secondary.example.com.
    $INCLUDE /etc/bind/local/inc.[% name %]